fix(solver): gate ParameterForCallExtractor's overload merge on noImplicitAny
tsc's `getIntersectedSignatures` returns `undefined` outright unless
`noImplicitAny` is on, so an overloaded callable contextual type never
contextually types a function expression's parameters under a non-strict
program. `ParameterExtractor::visit_callable` already carried this gate
(oracle-verified against `typescript@7.0.2`); the arity-filtered
`ParameterForCallExtractor::visit_callable` sibling, tried first for every
function-expression parameter, lacked it entirely and unconditionally
unioned disagreeing (and undeduped same-name generic) signature positions,
producing spurious TS2322 on `var f: { (x: string): string; (x: number):
number } = (x) => x;` and its generic-overload variants.

Split `ParameterForCallExtractor` into its own `extractors_for_call.rs` file
(architecture size ratchet on `contextual/extractors.rs` had no headroom for
the new field).
