+++ crates/tsz-solver/src/contextual/mod.rs
@@ -14,5 +14,6 @@
 
 mod core;
 pub(crate) mod extractors;
+pub(crate) mod extractors_for_call;
 
 pub use self::core::{ContextualTypeContext, apply_contextual_type, rest_argument_element_type};
